+++ app/src/main/java/com/example/coen390project/PH.java
@@ -5,19 +5,21 @@ import java.util.Date;
 
 public class PH {
     private String MEASUREMENT_DATE;
-    private Integer PH_VALUE;
+    private Float PH_VALUE;
 
-    public PH(Integer PH_VALUE, String MEASUREMENT_DATE){
+    public PH(Float PH_VALUE, String MEASUREMENT_DATE){
         this.MEASUREMENT_DATE = MEASUREMENT_DATE;
         this.PH_VALUE = PH_VALUE;
     }
 
-    public PH(Integer PH_VALUE){
+    public PH(Float PH_VALUE){
         this.PH_VALUE = PH_VALUE;
         SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd @ HH:mm:ss");
         String MEASUREMENT_DATE = sdf.format(new Date());
         setMEASUREMENT_DATE(MEASUREMENT_DATE);
     }
+
+
 
     public String getMEASUREMENT_DATE() {
         return MEASUREMENT_DATE;
@@ -27,11 +29,11 @@ public class PH {
         this.MEASUREMENT_DATE = MEASUREMENT_DATE;
     }
 
-    public Integer getPH_VALUE() {
+    public Float getPH_VALUE() {
         return PH_VALUE;
     }
 
-    public void setPH_VALUE(Integer PH_VALUE) {
+    public void setPH_VALUE(Float PH_VALUE) {
         this.PH_VALUE = PH_VALUE;
     }
 }
